app/requirements.txt
lxml==3.7.1
numpy==1.11.3
requests==2.12.4
PyYAML==3.12
flask==0.12
pyjwt==1.4.2
cryptography==1.3
